+++ CHANGELOG.rst
@@ -10,9 +10,15 @@ This change log uses principles from `keep a changelog <http://keepachangelog.co
 Added
 ^^^^^
 
+- ``dtoolcore.filehasher.hashsum_digest`` helper function
+- ``dtoolcore.filehasher.md5sum_digest`` helper function
+
 
 Changed
 ^^^^^^^
+
+- Improved name from ``dtoolcore.filehasher.hashsum`` to
+  ``dtoolcore.filehasher.hashsum_hexdigest``
 
 
 Deprecated
@@ -27,8 +33,10 @@ Fixed
 ^^^^^
 
 
+
 Security
 ^^^^^^^^
+
 
 [3.6.0] - 2018-08-03
 --------------------